refactor: drop the redundant prefix from the moved test files
Each test file now sits in the directory of the module it covers, so the
family name in `truncate/truncate_soundness_tests.rs` stutters. Rename
all five to `soundness_tests.rs` and update the one doc reference.
