+++ src/composer/range.rs
@@ -17,7 +17,7 @@ use super::{Composer, Constraint, WiredWitness, Witness};
 use crate::bit_iterator::BitIterator8;
 
 #[cfg(test)]
-mod range_soundness_tests;
+mod soundness_tests;
 
 /// Range-check gadgets
 impl Composer {
